# src/v3/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: API REST con Node JS
  description: Documentación de API REST con Node JS
  version: 1.0.0
  contact:
    name: JV Developer
    email: jv@email.com
servers:
  - url: http://{host}:{port}/api/{basePath}
    description: NodeJS
    variables:
      host:
        default: localhost
        description: Development server, Staging server, Production server example.com
        enum:
          - localhost
          - development
          - staging
          - api
      port:
        default: '3000'
        enum:
          - '3000'
          - '3001'
      basePath:
        default: v3
        enum:
          - v1
          - v2
          - v3
...
paths:
...
  /students/{id}:
    get:
      tags:
        - estudiantes
      summary: Buscar estudiante según su ID
      description: Buscar estudiante según su ID
      operationId: findStudentById
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: El id del estudiante
      responses:
        '200':
          description: Estudiante Encontrado Correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExitoListarEstudiantes'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
      security:
        - api_key: []
    put:
      tags:
        - estudiantes
      summary: Actualizar estudiante según su ID
      description: Actualizar datos del estudiante según su ID
      operationId: updateStudentById
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: El id del estudiante
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Student'
        required: true
      responses:
        '200':
          description: Estudiante Actualizado Correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExitoListarEstudiantes'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
      security:
        - api_key: []
    delete:
      tags:
        - estudiantes
      summary: Eliminar estudiante según su ID
      description: Eliminar datos del estudiante según su ID
      operationId: deleteStudentById
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: El id del estudiante
      responses:
        '200':
          description: Estudiante Eliminado Correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExitoListarEstudiantes'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
      security:
        - api_key: []
...
components:
  responses:
    BadRequest:
      description: >-
        (BadRequest) los datos enviados son incorrectos o hay datos obligatorios
        no enviados
    Unauthorized:
      description: (Unauthorized) no hay autorización para llamar el servicio
    NotFound:
      description: (NotFound) no se encontró la información
    ServerError:
      description: (ServerError) Error en el servidor
  schemas:
...
    Student:
      type: object
      properties:
        name:
          type: string
          description: nombre del estudiante
        age:
          type: integer
          description: edad del estudiante
        enroll:
          type: string
          description: estado del estudiante
      required:
        - name
        - age
        - enroll
      example:
        name: Kalef
        age: 10
        enroll: 'true'
...
    ExitoListarEstudiantes:
      type: object
      properties:
        id:
          type: integer
          enum:
            - 1
          description: el id del estudiante
        name:
          type: string
          enum:
            - gael
          description: nombre del estudiante
        age:
          type: integer
          enum:
            - 3
          description: edad del estudiante
        enroll:
          type: string
          enum:
            - 'true'
          description: estado del estudiante
...
  securitySchemes:
    api_key:
      type: apiKey
      name: Authorization
      in: header